+++ app/models/person.rb
@@ -1,11 +1,6 @@
 class Person < ApplicationRecord
-  # Attributes
-
-  # TODO: rename the DB column and remove all references to 'main'.
-  alias_attribute :owner, :main
-
   def companion?
-    !main?
+    !owner
   end
 
   def has_recent_recommendation?
@@ -63,7 +58,7 @@ class Person < ApplicationRecord
 
   NAME_MAX_LENGTH = 50
 
-  validates :account, uniqueness: { scope: :main }
+  validates :account, uniqueness: { scope: :owner }
 
   # Associations
 
@@ -82,6 +77,6 @@ class Person < ApplicationRecord
 
   # Scopes
 
-  scope :main,      -> { where(main: true) }
-  scope :companion, -> { where(main: false) }
+  scope :owner,     -> { where(owner: true) }
+  scope :companion, -> { where(owner: false) }
 end